Fixed Debug Print(int)

+++ waypointNavigator/waypointNavigator/debug.h
@@ -8,6 +8,7 @@
 #include "usart.h"
 #include "main.h"
 #include <stdio.h>
+#include <stdlib.h>
 
 void debug_print(const char* debugString);
-void debug_print(double debugNum);
+void debug_print(int debugNum);
